[angular-scroll] Fix module exports
