+++ types/angular-scroll/angular-scroll-tests.ts
@@ -1,4 +1,5 @@
-
+import * as angular from 'angular';
+import duScroll = require('angular-scroll');
 
 namespace TestApp {
 
@@ -59,6 +60,6 @@ namespace TestApp {
         }
     }
 
-    angular.module('testApp', ['duScroll'])
+    angular.module('testApp', [duScroll])
         .controller('testController', TestController);
 }
